gemini/pyproject.toml
[project]
name = "gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.36.0",
    "google-generativeai>=0.7.0",
]
